+++ Bsp/bsp_beep.h
@@ -9,6 +9,6 @@ void bsp_beep_init(void);
 void bsp_beep_on(void);
 void bsp_beep_off(void);
 void bsp_beep_toggle(void);
-void bsp_beep(uint32_t time, uint32_t count);
+void bsp_beep(uint16_t time, uint16_t count);
 
 #endif  // __BSP_BEEP_H__
